+++ src/Config/ConfigFunction/ConfigFunction.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace YaFou\Mapik\Config\ConfigFunction;
+
+interface ConfigFunction
+{
+    public function call(array $arguments): string;
+}