+++ src/Exception/ConfigException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace YaFou\Mapik\Exception;
+
+use Exception;
+
+class ConfigException extends Exception
+{
+    
+}